Fixed the longest-standing bug on the site
So, there's this bug that's plagued the site since its inception. iOS
Safari eventually seems to have a memory leak and starts cutting off
all transitions before they've reached their end. It gets
progressively worse until quitting the app is required. It can be
triggered by drag-and-dropping elements repeatedly *anywhere* in
Safari, and affects all webpages with CSS transitions.

In iOS 13.4, it seems Apple has miraculously fixed this nightmare. But
for whatever reason, AOS is still problematic. If an element has a
nonzero delay, it will be bugged, but zero-delay elements behave as
usual. And so the solution is, unfortunately, to handle almost all of
what AOS does manually.